+++ app/(auth)/register.js
@@ -13,7 +13,6 @@ import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { styled } from 'nativewind';
 import { Button } from '../../components/Buttons';
-import { createTutorial } from '../../backend/firebaseFunctions';
 import { passwordValidation } from '../../backend/functions';
 import { BottomSheetModal, BottomSheetFlatList, BottomSheetModalProvider } from '@gorhom/bottom-sheet';
 import { handle, backdrop, SnapPoints } from '../../components/BottomSheetModalHelpers.js';
@@ -374,23 +373,6 @@ async function createUserData(
 	image,
 	authContext
 ) {
-	let data = {
-		title: 'Tutorial',
-		description: 'A place to learn how to use Prayer Circle',
-		iconColor: '000000',
-		icon: 'home',
-		timestamp: Date.now(),
-		type: 'individual',
-		color: 'FFFFFF',
-		members: {},
-		admin: {},
-		usersAwaitingEntry: {},
-		owner: false
-		// this data structure is taken from createCircle.js, but owner seems to have its type changed when its passed to createCircle in firebaseFunctions; am I reading that wrong?
-	};
-
-	createTutorial(data);
-
 	if (fname.length < 1 || lname.length < 1)
 		return notify('Registration Error', 'Invalid Name', '#CC2500'); // check name length
 
@@ -432,16 +414,6 @@ async function createUserData(
 		private: {
 			email: email,
 			timestamp: Date.now(),
-			circles: {
-				'-NiN-27IuGR02mcGS2CS': {
-					permissions: {
-						admin: false,
-						owner: false,
-						read: true,
-						write: true
-					}
-				}
-			},
 			reactions: false,
 			comments: false,
 			posts: false,
